fix(core): store integer fields as bigint (#2425)
* fix(core): widen integer storage to bigint

* fix(core): preserve numeric defaults and json bigint ddl

+++ packages/core/src/migrations/differ.ts
@@ -1875,12 +1875,12 @@ export class SchemaComparer {
           manifestNormalized === 'INTEGER' &&
           dbNormalized === 'TEXT'
         ) {
-          typeClause += ` USING trim(${quotedCol}::text)::integer`;
+          typeClause += ` USING trim(${quotedCol}::text)::bigint`;
         } else if (
           manifestNormalized === 'INTEGER' &&
           dbNormalized === 'REAL'
         ) {
-          typeClause += ` USING ${quotedCol}::integer`;
+          typeClause += ` USING ${quotedCol}::bigint`;
         } else if (
           manifestNormalized === 'TIMESTAMP' &&
           (dbNormalized === 'TEXT' || dbNormalized === 'JSON')
